refactor: rename `NetworkCustomPeer` to `NetworkCustomPeerModal`
